+++ src/cli/main.go
@@ -3,154 +3,19 @@ package main
 import (
   "fmt"
   "log"
-  "io/ioutil"
 
   "github.com/jroimartin/gocui"
-  "gopkg.in/yaml.v2"
 )
 
-// TYPES
+// CONSTANTS
 
-type state struct {
-  cfg map[string]interface{}
-  option option
-}
+const VIEW_MAIN = "main"
+const VIEW_SIDE = "side"
+const VIEW_INPUT = "input"
+const VIEW_SAVE = "save"
+const VIEW_CONSOLE = "console"
 
-func (s state) Copy() state {
-  newState := s
-  newState.cfg = make(map[string]interface{})
-  for k,v := range s.cfg {
-    newState.cfg[k] = v
-  }
-  return newState
-}
-
-type option interface {
-  Present(g *gocui.Gui, v *gocui.View) error
-  Name() string
-  IsModuleConfig() bool
-}
-
-type multiOption struct {
-  options []string
-  name string
-  isModuleConfig bool
-}
-
-func (mo multiOption) Present(g *gocui.Gui, v *gocui.View) error {
-  for _, option := range mo.options {
-    fmt.Fprintln(v, option)
-  }
-  v.Highlight = true
-  if _, err := g.SetCurrentView("main"); err != nil {
-    return err
-  }
-  return nil
-}
-
-func (mo multiOption) Name() string {
-  return mo.name
-}
-
-func (mo multiOption) IsModuleConfig() bool {
-  return mo.isModuleConfig
-}
-
-type textInputOption struct {
-  prompt string
-  name string
-  isModuleConfig bool
-}
-
-func (to textInputOption) Present(g *gocui.Gui, v *gocui.View) error {
-  fmt.Fprintln(v, to.prompt)
-  v.Highlight = false
-  maxX, maxY := g.Size()
-  if iView, err := g.SetView("input", maxX/2-30, maxY/2, maxX/2+30, maxY/2+2); err != nil {
-    if err != gocui.ErrUnknownView {
-      return err
-    }
-    iView.Editable = true
-    if _, err := g.SetCurrentView("input"); err != nil {
-      return err
-    }
-  }
-  return nil
-}
-
-func (to textInputOption) Name() string {
-  return to.name
-}
-
-func (to textInputOption) IsModuleConfig() bool {
-  return to.isModuleConfig
-}
-
-type saveOption struct {}
-
-func (so saveOption) Present(g *gocui.Gui, v *gocui.View) error {
-
-  fmt.Fprintln(v, "SAVE CONFIG")
-  fmt.Fprintln(v, "-----------")
-  fmt.Fprintln(v, "")
-  fmt.Fprintln(v, "Please choose a file name for this workflow\nconfiguration.\n\nIt will be saved as a yaml file in your workflows\ndirectory.")
-
-  maxX, maxY := g.Size()
-  if iView, err := g.SetView("save", maxX/2-30, maxY/2, maxX/2+30, maxY/2+2); err != nil {
-    if err != gocui.ErrUnknownView {
-      return err
-    }
-    iView.Editable = true
-    if _, err := g.SetCurrentView("save"); err != nil {
-      return err
-    }
-  }
-  return nil
-}
-
-func (to saveOption) Name() string {
-  return "save"
-}
-
-func (to saveOption) IsModuleConfig() bool {
-  return false
-}
-
-
-// GLOBAL STATE
-
-var history []state = []state {}
-
-// MUTATIONS
-
-func pushState(g *gocui.Gui, newState state) {
-  history = append(history, newState)
-  updateConfigView(g, newState.cfg)
-  updateOptionView(g, newState.option)
-}
-
-func popState(g *gocui.Gui) {
-  if len(history) > 1 {
-    history = history[:len(history)-1]
-    updateConfigView(g, history[len(history)-1].cfg)
-    updateOptionView(g, history[len(history)-1].option)
-  }
-}
-
-func update(g *gocui.Gui, key string, value string, isModuleConfig bool) {
-  newState := history[len(history)-1].Copy()
-  if !isModuleConfig {
-      newState.cfg[key] = value
-  } else {
-      if _, ok := newState.cfg["config"]; !ok {
-        newState.cfg["config"] = make(map[string]string)
-      }
-      config := newState.cfg["config"].(map[string]string)
-      config[key] = value
-  }
-  newState.option = getNextOption(g)
-  pushState(g, newState)
-}
+const DIR_WORKFLOWS = "../../workflows"
 
 // GOCUI LIFECYCLE
 
@@ -182,13 +47,13 @@ func layout(g *gocui.Gui) error {
   helpHeight := 4
   logHeight := 30
 
-  if v, err := g.SetView("main", 0, 0, sideWidth-2, maxY-helpHeight-1); err != nil {
+  if v, err := g.SetView(VIEW_MAIN, 0, 0, sideWidth-2, maxY-helpHeight-1); err != nil {
     if err != gocui.ErrUnknownView {
       return err
     }
     v.SelBgColor = gocui.ColorGreen
     v.SelFgColor = gocui.ColorBlack
-    if _, err := g.SetCurrentView("main"); err != nil {
+    if _, err := g.SetCurrentView(VIEW_MAIN); err != nil {
       return err
     }
     v.Wrap = true
@@ -204,13 +69,13 @@ func layout(g *gocui.Gui) error {
     fmt.Fprintln(v, " Quit: ctrl+c")
   }
 
-  if _, err := g.SetView("side", sideWidth, 0, maxX-1, maxY-logHeight-1); err != nil {
+  if _, err := g.SetView(VIEW_SIDE, sideWidth, 0, maxX-1, maxY-logHeight-1); err != nil {
     if err != gocui.ErrUnknownView {
       return err
     }
   }
 
-  if v, err := g.SetView("log", sideWidth, maxY - logHeight, maxX-1, maxY-1); err != nil {
+  if v, err := g.SetView(VIEW_CONSOLE, sideWidth, maxY - logHeight, maxX-1, maxY-1); err != nil {
     if err != gocui.ErrUnknownView {
       return err
     }
@@ -233,19 +98,19 @@ func keybindings(g *gocui.Gui) error {
   if err := g.SetKeybinding("", gocui.KeyCtrlZ, gocui.ModNone, undo); err != nil {
     log.Panicln(err)
   }
-  if err := g.SetKeybinding("main", gocui.KeyArrowUp, gocui.ModNone, cursorUp); err != nil {
+  if err := g.SetKeybinding(VIEW_MAIN, gocui.KeyArrowUp, gocui.ModNone, cursorUp); err != nil {
     return err
   }
-  if err := g.SetKeybinding("main", gocui.KeyArrowDown, gocui.ModNone, cursorDown); err != nil {
+  if err := g.SetKeybinding(VIEW_MAIN, gocui.KeyArrowDown, gocui.ModNone, cursorDown); err != nil {
     return err
   }
-  if err := g.SetKeybinding("main", gocui.KeyEnter, gocui.ModNone, selectOption); err != nil {
+  if err := g.SetKeybinding(VIEW_MAIN, gocui.KeyEnter, gocui.ModNone, selectOption); err != nil {
     return err
   }
-  if err := g.SetKeybinding("input", gocui.KeyEnter, gocui.ModNone, inputEntered); err != nil {
+  if err := g.SetKeybinding(VIEW_INPUT, gocui.KeyEnter, gocui.ModNone, inputEntered); err != nil {
     return err
   }
-  if err := g.SetKeybinding("save", gocui.KeyEnter, gocui.ModNone, saveFile); err != nil {
+  if err := g.SetKeybinding(VIEW_SAVE, gocui.KeyEnter, gocui.ModNone, saveFile); err != nil {
     return err
   }
 
@@ -260,7 +125,7 @@ func saveFile(g *gocui.Gui, v *gocui.View) error {
     return nil
   }
   input = input[:len(input)-1]  // remove trailing newline
-  path := "../../workflows/" + input + ".yaml"
+  path := DIR_WORKFLOWS + "/" + input + ".yaml"
   data := history[len(history)-1].cfg
   writeToYamlFile(path, data)
   return gocui.ErrQuit
@@ -277,11 +142,11 @@ func inputEntered(g *gocui.Gui, v *gocui.View) error {
   isModuleConfig := history[len(history)-1].option.IsModuleConfig()
   update(g, optionName, input, isModuleConfig)
 
-  if err := g.DeleteView("input"); err != nil {
+  if err := g.DeleteView(VIEW_INPUT); err != nil {
     // log.Panicln("here")
     return nil
   }
-  if _, err := g.SetCurrentView("main"); err != nil {
+  if _, err := g.SetCurrentView(VIEW_MAIN); err != nil {
     return err
   }
 
@@ -361,7 +226,7 @@ func printToView(g *gocui.Gui, view string, text string) error {
 
 func updateConfigView(g *gocui.Gui, cfg map[string]interface{}) error {
 
-    v, err := g.View("side")
+    v, err := g.View(VIEW_SIDE)
     if err != nil {
       return err
     }
@@ -385,21 +250,21 @@ func updateConfigView(g *gocui.Gui, cfg map[string]interface{}) error {
 func updateOptionView(g *gocui.Gui, o option) error {
   logger(g, "update option")
 
-  v, err := g.View("main")
+  v, err := g.View(VIEW_MAIN)
   if err != nil {
     return err
   }
   if v != nil {
     v.Clear()
-    if err := g.DeleteView("input"); err != nil {}
-    if err := g.DeleteView("save"); err != nil {}
+    if err := g.DeleteView(VIEW_INPUT); err != nil {}
+    if err := g.DeleteView(VIEW_SAVE); err != nil {}
     o.Present(g, v)
   }
   return nil
 }
 
 func logger(g *gocui.Gui, text string) {
-  v, _ := g.View("log")
+  v, _ := g.View(VIEW_CONSOLE)
   if v != nil {
     fmt.Fprintln(v, text)
   }
@@ -414,16 +279,5 @@ func getNextOption(g *gocui.Gui) option {
     return textInputOption{ prompt: "please enter a folder path", name: "folder", isModuleConfig: false  }
   } else {
     return saveOption{}
-  }
-}
-
-func writeToYamlFile(path string, data map[string]interface{}) {
-  d, err := yaml.Marshal(&data)
-  if err != nil {
-          log.Fatalf("error: %v", err)
-  }
-  err = ioutil.WriteFile(path, d, 0644)
-  if err != nil {
-    panic(err)
   }
 }
